Split into Controller and Service
