+++ spring-boot/admin/src/main/java/eu/coatrack/admin/controllers/GatewayApiController.java
@@ -23,10 +23,14 @@ package eu.coatrack.admin.controllers;
 import eu.coatrack.admin.model.repository.ApiKeyRepository;
 import eu.coatrack.admin.model.repository.ProxyRepository;
 import eu.coatrack.admin.model.repository.ServiceApiRepository;
+import eu.coatrack.admin.service.GatewayApiService;
 import eu.coatrack.api.ApiKey;
 import eu.coatrack.api.Proxy;
 import eu.coatrack.api.ServiceApi;
 import javassist.NotFoundException;
+import lombok.extern.log4j.Log4j;
+import lombok.extern.log4j.Log4j2;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,62 +48,27 @@ import java.util.stream.Collectors;
  * Controller that handles HTTP calls by CoatRack gateways.
  */
 
+@Slf4j
 @Controller
 public class GatewayApiController {
-
-    private static final Logger log = LoggerFactory.getLogger(GatewayApiController.class);
-
     @Autowired
-    ApiKeyRepository apiKeyRepository;
-
-    @Autowired
-    ServiceApiRepository serviceApiRepository;
-
-    @Autowired
-    ProxyRepository proxyRepository;
+    private GatewayApiService gatewayApiService;
 
     @RequestMapping(value = "/api/api-keys/search/findByKeyValue", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
     public ApiKey findApiKeyEntityByApiKeyValue(@RequestParam("keyValue") String apiKeyValue) {
-        return apiKeyRepository.findByKeyValue(apiKeyValue);
+        return gatewayApiService.findApiKeyEntityByApiKeyValue(apiKeyValue);
     }
 
     @RequestMapping(value = "/api/services/search/findByApiKeyValue", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
     public ServiceApi findServiceByApiKeyValue(@RequestParam("apiKeyValue") String apiKeyValue) {
-        return serviceApiRepository.findByApiKeyValue(apiKeyValue);
+        return gatewayApiService.findServiceByApiKeyValue(apiKeyValue);
     }
 
     //Due to legacy code reasons the gateway API key and the gateway id are exactly the same.
     @GetMapping("/api/gateways/api-keys")
     public ResponseEntity<List<ApiKey>> findApiKeyListByGatewayApiKey(@RequestParam("gateway-api-key") String gatewayIdAndApiKey) {
-        log.debug("The gateway with the ID {} requests its latest API key list.", gatewayIdAndApiKey);
-        try {
-            Optional<Proxy> callingProxy = proxyRepository.findById(gatewayIdAndApiKey);
-
-            if (callingProxy.isPresent()) {
-                Proxy proxyToUpdate = callingProxy.get();
-                proxyToUpdate.updateTimeOfLastSuccessfulCallToAdmin_setToNow();
-                proxyRepository.save(proxyToUpdate);
-                return new ResponseEntity<>(getApiKeysBelongingToServicesOf(proxyToUpdate), HttpStatus.OK);
-            } else {
-                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-            }
-
-        } catch (Exception e) {
-            log.warn("The creation of the API key list for the gateway {} failed.", gatewayIdAndApiKey, e);
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    private List<ApiKey> getApiKeysBelongingToServicesOf(Proxy proxy) {
-        if (proxy.getServiceApis() == null || proxy.getServiceApis().isEmpty()) {
-            log.debug("The gateway with the ID {} does not provide any services.", proxy.getId());
-            return new ArrayList<>();
-        } else
-            return proxy.getServiceApis().stream()
-                    .flatMap(
-                            serviceApi -> serviceApi.getApiKeys().stream()
-                    ).collect(Collectors.toList());
+        return gatewayApiService.findApiKeyListByGatewayApiKey(gatewayIdAndApiKey);
     }
 }
